# PERF: improve DTI string parse