@jreback 2016-07-18
@sinhrks this cleans up lots of junk!

just a couple of clarifications.